Add note to explain `update_all` example
`Ecto.Repo.update_all/3` can accept updates either in second argument (`updates`) and in its first argument (`queryable`) if the query has an `update:` key. While this is shown in the examples, it isn't described in the text, and it took me a while and a lot of trial and error to figure out exactly what was going on. (Specifically, it wasn't clear that I could pass `[]` for the `updates` argument and it would still work.)

This PR adds a small note to the `update_all` doc. From reading the ExDoc docs I don't know if comments are allowed in the examples; if not, let me know and I can move the text elsewhere (if the change is acceptable).

Thanks for your great work!